+++ test/util/pb_graph_test.c
@@ -1,4 +1,4 @@
-#include <libcompat.h>
+#include "../test_util.h"
 #include <check.h>
 #include <pb/util/pb_graph.h>
 #include <pb/util/pb_hash_utils.h>
@@ -29,22 +29,22 @@ void pb_graph_test_teardown() {
 
 START_TEST(add_vertex)
 {
-    pb_graph_add_vertex(graph, (void*)1, (void*)1);
+    pb_graph_add_vertex(graph, (void*)((size_t)1), (void*)((size_t)1));
     ck_assert_msg(graph->vertices->size == 1, "Graph didn't contain any vertices after pb_graph_add_vertex");
 }
 END_TEST
 
 START_TEST(get_vertex)
 {
-    pb_vertex const* vert = pb_graph_get_vertex(graph, (void*)1);
+    pb_vertex const* vert = pb_graph_get_vertex(graph, (void*)((size_t)1));
     ck_assert_msg(vert != NULL, "graph didn't contain vertex.");
-    ck_assert_msg(vert->data == (void*)1, "vertex data should have been 1, was %p", vert->data);
+    ck_assert_msg(vert->data == (void*)((size_t)1), "vertex data should have been 1, was %p", vert->data);
 }
 END_TEST
 
 START_TEST(get_nonexistent_vertex)
 {
-    pb_vertex const* vert = pb_graph_get_vertex(graph, (void*)3);
+    pb_vertex const* vert = pb_graph_get_vertex(graph, (void*)((size_t)3));
     ck_assert_msg(vert == NULL, "vert should have been NULL, was %p", vert);
 }
 END_TEST
@@ -53,7 +53,7 @@ START_TEST(remove_vertex)
 {
     /* Remove the vertex we added in the previous test */
     /* Unit tests aren't supposed to depend on order, but this is more for convenience; I could create a new graph, add to it, and remove stuff just as well */
-    pb_graph_remove_vertex(graph, (void*)1);
+    pb_graph_remove_vertex(graph, (void*)((size_t)1));
 
 	ck_assert_msg(graph->vertices->size == 0, "Graph size was %lu, should have been 0.", graph->vertices->size);
 }
@@ -65,24 +65,24 @@ START_TEST(add_edge)
     pb_vertex const* v0;
     pb_vertex const* v1;
 
-    result = pb_graph_add_vertex(graph, (void*)0, (void*)0);
+    result = pb_graph_add_vertex(graph, (void*)((size_t)0), (void*)((size_t)0));
     if (result == -1) {
         ck_abort_msg("Ran out of memory while adding vertex.");
     }
 
-    result = pb_graph_add_vertex(graph, (void*)1, (void*)1);
+    result = pb_graph_add_vertex(graph, (void*)((size_t)1), (void*)((size_t)1));
     if (result == -1) {
         ck_assert_msg("Ran out of memory while adding vertex.");
     }
 
-	result = pb_graph_add_edge(graph, (void*)0, (void*)1, 0.0f, (void*)5);
+	result = pb_graph_add_edge(graph, (void*)((size_t)0), (void*)((size_t)1), 0.0f, (void*)((size_t)5));
 	if (result == -1) {
 		/* Somehow we ran out of memory during the test. Oops */
 		ck_abort_msg("Unexpectedly out of memory during test.");
 	}
 
-    v0 = pb_graph_get_vertex(graph, (void*)0);
-    v1 = pb_graph_get_vertex(graph, (void*)1);
+    v0 = pb_graph_get_vertex(graph, (void*)((size_t)0));
+    v1 = pb_graph_get_vertex(graph, (void*)((size_t)1));
 
     ck_assert_msg(v0->edges_size == 1, "No edge was added to v0.");
 }
@@ -94,9 +94,9 @@ START_TEST(get_edge)
     pb_vertex const* v1;
     pb_edge const* edge;
 
-    v0 = pb_graph_get_vertex(graph, (void*)0);
-    v1 = pb_graph_get_vertex(graph, (void*)1);
-    edge = pb_graph_get_edge(graph, (void*)0, (void*)1);
+    v0 = pb_graph_get_vertex(graph, (void*)((size_t)0));
+    v1 = pb_graph_get_vertex(graph, (void*)((size_t)1));
+    edge = pb_graph_get_edge(graph, (void*)((size_t)0), (void*)((size_t)1));
 
     ck_assert_msg(edge != NULL, "graph didn't contain edge from v0 to v1");
     ck_assert_msg(v0->edges[0] == edge, "v0 doesn't contain correct edge.");
@@ -107,15 +107,15 @@ END_TEST
 
 START_TEST(get_nonexistent_edge)
 {
-    pb_edge const* edge = pb_graph_get_edge(graph, (void*)5, (void*)6);
+    pb_edge const* edge = pb_graph_get_edge(graph, (void*)((size_t)5), (void*)((size_t)6));
     ck_assert_msg(edge == NULL, "Incorrect return value (should have been NULL, was %p).", edge);
 }
 END_TEST
 
 START_TEST(remove_edge)
 {
-    int result = pb_graph_remove_edge(graph, (void*)0, (void*)1);
-    pb_vertex const* v0 = pb_graph_get_vertex(graph, (void*)0);
+    int result = pb_graph_remove_edge(graph, (void*)((size_t)0), (void*)((size_t)1));
+    pb_vertex const* v0 = pb_graph_get_vertex(graph, (void*)((size_t)0));
 
     ck_assert_msg(result == 0, "Incorrect return value (should have been 0, was %d).", result);
     ck_assert_msg(graph->edges->size == 0, "graph->edges->size should have been 0, was %lu", graph->edges->size);
@@ -125,12 +125,12 @@ END_TEST
 
 START_TEST(remove_vertex_with_edge)
 {
-    pb_vertex const* v1 = pb_graph_get_vertex(graph, (void*)1);
+    pb_vertex const* v1 = pb_graph_get_vertex(graph, (void*)((size_t)1));
 
-    pb_graph_add_edge(graph, (void*)0, (void*)1, 0.f, (void*)10);
-    pb_graph_add_edge(graph, (void*)1, (void*)0, 1.f, (void*)3);
+    pb_graph_add_edge(graph, (void*)((size_t)0), (void*)((size_t)1), 0.f, (void*)((size_t)10));
+    pb_graph_add_edge(graph, (void*)((size_t)1), (void*)((size_t)0), 1.f, (void*)((size_t)3));
 
-    pb_graph_remove_vertex(graph, (void*)0);
+    pb_graph_remove_vertex(graph, (void*)((size_t)0));
     
     ck_assert_msg(graph->edges->size == 0, "graph->edges->size should have been 0, was %lu", graph->edges->size);
     ck_assert_msg(v1->edges_size == 0, "v1->edges_size should have been 0, was %lu", v1->edges_size);
@@ -145,8 +145,8 @@ static void add_one_to_vertex_data(void const* vert_id, pb_vertex* vert, void* p
 START_TEST(graph_for_each_vertex)
 {
     pb_graph* g = pb_graph_create(pb_pointer_hash, pb_pointer_eq);
-    int items[] = { 1, 2, 3 };
-    int expected[] = { 2, 3, 4 };
+    size_t items[] = { 1, 2, 3 };
+    size_t expected[] = { 2, 3, 4 };
     int i;
 
     /* Add a pointer to each element in items as a vertex */
@@ -172,11 +172,11 @@ static void add_one_to_edge_data(pb_edge const* edge, void* param) {
 START_TEST(graph_for_each_edge)
 {
     pb_graph* g = pb_graph_create(pb_pointer_hash, pb_pointer_eq);
-    int items[] = { 1, 2, 3 };
-    int expected[] = { 2, 3, 4 };
-    int i;
+    size_t items[] = { 1, 2, 3 };
+    size_t expected[] = { 2, 3, 4 };
+    size_t i;
 
-    int vert_items[] = { 5, 6, 7, 8 };
+    size_t vert_items[] = { 5, 6, 7, 8 };
 
     /* Add a pointer to each element in vert_items as a vertex */
     for (i = 0; i < 4; ++i) {
